requirements.txt
# Core pipeline
pipecat-ai[websocket,deepgram,kokoro,silero,openai]==0.0.106

# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.3

# Database
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2

# Object storage
minio==7.2.15

# Config
python-dotenv==1.1.0
pydantic-settings==2.9.1
